agrega control en people_options
