+++ src/services/people/checkUnverified.js
@@ -1,0 +1,51 @@
+const { PEOPLE_STATE_UNVERIFIED, PEOPLE_STATE_ACTIVE, USER_PROVIDER } = require('../../constants');
+const { People, People_options, Categories_options, Categories } = require('../../db');
+
+const checkUnverified = async (idPeople) => {
+    try {
+        //revisar state por si hay que pasarlo a activo o unverified
+        const people = await People.findByPk(idPeople)
+        if (people.dataValues.typeOfPerson === USER_PROVIDER &&
+            (people.dataValues.state === PEOPLE_STATE_UNVERIFIED || people.dataValues.state === PEOPLE_STATE_ACTIVE)) {
+            //reviso que tenga servicios
+            const { count: countServices, rows } = await People_options.findAndCountAll({
+                where: {
+                    idPeople: people.dataValues.idPeople,
+                    isDeleted: false
+                },
+                include: [
+                    {
+                        model: Categories_options,
+                        required: true,
+                        include: [
+                            {
+                                model: Categories,
+                                required: true,
+                                where: {
+                                    isService: true
+                                },
+                            }
+                        ]
+                    }
+                ]
+            });
+            //reviso que tenga calendario
+            const hasCalendar = people.dataValues.weekCalendar.find(value => value === true);
+            //
+            if (people.dataValues.idLocation &&
+                people.dataValues.geoposition &&
+                people.dataValues.image &&
+                people.dataValues.phone &&
+                countServices > 0 &&
+                hasCalendar) {
+                    await people.update({state:PEOPLE_STATE_ACTIVE})
+            }else{
+                await people.update({state:PEOPLE_STATE_UNVERIFIED})
+            }
+        }
+    } catch (error) {
+        throw error;
+    }
+};
+
+module.exports = { checkUnverified };
